+++ src/ts/export.ts
@@ -23,6 +23,9 @@ export class Export {
     if (!Is.object(data[0]) && !Is.json(data[0])) {
       return false;
     }
+    if (!filename.trim().length) {
+      filename = 'export';
+    }
     if (options) {
       this.setOptions(options);
     }